Layer1-2/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
numpy==1.26.4
PyWavelets==1.6.0
imagehash==4.3.1
librosa==0.10.2
Pillow==10.3.0
pydantic==2.7.4
soundfile==0.12.1
httpx==0.27.0
